app/pyproject.toml
[tool.poetry]
name = "fastapi-structlog"
version = "0.1.0"
description = ""
authors = ["Shinya Matsuura <“shinya.matsuura@hennge.com”>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.59.0"
uvicorn = "^0.11.5"
structlog = "^20.1.0"
gunicorn = "^20.0.4"
python-rapidjson = "^0.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
